Fixed warnings from x64 cfinput.cpp
Now, no warnings from VS2013x64nc
and no warnings from VS2010x32nc.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/tech_fixes@8212 70805675-57b6-444b-bb00-18ecffcb4be8

+++ modules/cfinput.cpp
@@ -806,10 +806,10 @@ bool CFInput::getclimate(Gridcell& gridcell) {
 
 			int years_to_simulate = nyear_spinup + historic_years;
 
-			int cells_done = distance(gridlist.begin(), current_gridcell);
+			int cells_done = (int)distance(gridlist.begin(), current_gridcell);
 
-			double progress=(double)(cells_done*years_to_simulate+date.year)/
-				(double)(gridlist.size()*years_to_simulate);
+			double progress = (double)(cells_done*years_to_simulate+date.year)/
+				(gridlist.size()*(double)years_to_simulate);
 			tprogress.setprogress(progress);
 			dprintf("%3d%% complete, %s elapsed, %s remaining\n",(int)(progress*100.0),
 				tprogress.elapsed.str,tprogress.remaining.str);
